finish authoz

+++ controller/PermissionController.php
@@ -1,13 +1,15 @@
 <?php
 require_once ('models/Permission.php');
-
+require_once ('controller/CheckPermissionController.php');
 class PermissionController{
 
     public $permission;
+    public $checkPermission;
 
     public function __construct()
     {
         $this->permission = new Permission();
+        $this->checkPermission = new CheckPermissionController();
 
     }
 
@@ -20,57 +22,108 @@ class PermissionController{
 
     public function add_permission()
     {
-        require_once('views/admins/permission/add.php');
+        $per = 'add permission';
+        $check =  $this->checkPermission->CheckPer($per);
+
+        if($check){
+            require_once('views/admins/permission/add.php');
+        }else{
+            header('Location: ?view=admin&act=403');
+        }
+
+
+
     }
 
     public function store_permission()
     {
-        date_default_timezone_set('Asia/Ho_Chi_Minh');
+        $per = 'add permission';
+        $check =  $this->checkPermission->CheckPer($per);
+        if($check){
+            date_default_timezone_set('Asia/Ho_Chi_Minh');
 
-        $data = [
+            $data = [
 
-            'name' => $_POST['name'],
-            'created_at' => date("Y-m-d H:i:s"),
+                'name' => $_POST['name'],
+                'created_at' => date("Y-m-d H:i:s"),
 
-        ];
+            ];
 
-        $new_permission = $this->permission->insert($data);
-        header("Location: ?view=admin&act=list_permission");
+            $new_permission = $this->permission->insert($data);
+            header("Location: ?view=admin&act=list_permission");
+        }else{
+            header('Location: ?view=admin&act=403');
+        }
+
+
+
     }
 
     public function edit_permission()
     {
-        $id = $_GET['id'];
-        $data = $this->permission->edit($id);
-        require_once('views/admins/permission/edit.php');
+        $per = 'edit permission';
+        $check =  $this->checkPermission->CheckPer($per);
+
+        if($check){
+
+            $id = $_GET['id'];
+            $data = $this->permission->edit($id);
+            require_once('views/admins/permission/edit.php');
+        }else{
+            header('Location: ?view=admin&act=403');
+        }
+
+
+
     }
 
     public function update_permission()
     {
-        date_default_timezone_set('Asia/Ho_Chi_Minh');
+        $per = 'edit permission';
+        $check =  $this->checkPermission->CheckPer($per);
+        if($check){
+            date_default_timezone_set('Asia/Ho_Chi_Minh');
 
-        $id = $_GET['id'];
+            $id = $_GET['id'];
 
-        $data = [
+            $data = [
 
-            'name' => $_POST['name'],
-            'created_at' => date("Y-m-d H:i:s"),
-            'updated_at' => date("Y-m-d H:i:s"),
+                'name' => $_POST['name'],
+                'created_at' => date("Y-m-d H:i:s"),
+                'updated_at' => date("Y-m-d H:i:s"),
 
-        ];
+            ];
 
-        $permission = $this->permission->updatePermission($data, $id);
-        header('Location: index.php?view=admin&act=list_permission');
+            $permission = $this->permission->updatePermission($data, $id);
+            header('Location: index.php?view=admin&act=list_permission');
+        }else{
+            header('Location: ?view=admin&act=403');
+        }
+
+
 
     }
 
     public function delete_permission()
     {
-        $id = $_GET['id'];
-        $this->permission->delete($id);
-        header('Location: ?view=admin&act=list_permission');
+        $per = 'delete permission';
+        $check =  $this->checkPermission->CheckPer($per);
+        if($check){
+            $id = $_GET['id'];
+            $this->permission->delete($id);
+            header('Location: ?view=admin&act=list_permission');
+        }else{
+            header('Location: ?view=admin&act=403');
+        }
+
 
     }
+
+    function error403(){
+        require_once('views/admins/403.php');
+    }
+
+
